Regex check coordinates
I switched the coordinate check around to use a regular expression
rather than just a string split.

That way the player can use any display they want for grid numbers,
as long as it is "h" and two numbers.

I also switch the array indices around to accomodate for the regular
expression change.